drm/msm/dsi: split bridge pre_enable/post_disable
In DSI video mode the video stream should not be active yet when the
panel pre_enable() hook is called. However this is impossible currently
because the DSI host is powered up and started in the same bridge
pre_enable() hook. Split this into pre_enable() and enable() and start
the video stream only during enable(), after the panel pre_enable()
has already run. Same for disable()/post_disable().

Fixes: 007ac0262b0d ("drm/msm/dsi: switch to DRM_PANEL_BRIDGE")
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>